audio_agent_v1/requirements.txt
anthropic==0.87.0
beautifulsoup4==4.14.3
faiss-cpu==1.13.0
httpx==0.28.1
numpy==2.0.2
playwright==1.58.0
python-dotenv==1.2.1
sentence-transformers==5.1.2
torch==2.8.0
transformers==4.57.6
